assistant-service/requirements.txt
requests
playwright
pinecone-client
google-generativeai